feat: Cập nhật hệ thống chest, enum ArenaVariant và các tối ưu hóa liên quan ✨ Cải tiến chính:
Chuyển đổi ArenaVariant thành enum đầy đủ: Thay thế class ArenaVariant bằng enum với các giá trị cụ thể (SOLO_NORMAL, SOLO_INSANE, v.v.), thêm phương thức fromKey để dễ dàng lấy variant từ mode và type. Cập nhật getter cho mode và type, đồng thời cải thiện xử lý lỗi khi parse key.
Tối ưu hóa xử lý chest items:

Thay đổi cấu trúc map chestItems từ Map<ArenaState, Map<ArenaVariant, ...>> sang Map<ArenaVariant, Map<ArenaState, ...>> để ưu tiên variant làm key chính, giúp dễ quản lý hơn theo biến thể game.
Cập nhật phương thức loadChestItems: Parse từ file refill.yml theo cấu trúc mới, bỏ hỗ trợ inherit (tạm thời), và tập trung vào variant làm gốc.
Nâng cấp parseChestItem: Chuyển chance từ double sang int, cải thiện parsing với split linh hoạt hơn, hỗ trợ enchant, flag, lore, name chi tiết, và xử lý lỗi tốt hơn (warning cho material không hợp lệ).

Cập nhật pickRandomItems: Sử dụng ThreadLocalRandom cho ngẫu nhiên hiệu quả, thêm logic cơ hội (Utils.chanceOf), giới hạn attempts để tránh lặp vô hạn, và fallback chọn ngẫu nhiên nếu hết attempts.

🔧 Sửa lỗi và tinh chỉnh:

Skywars.java: Thêm import cần thiết (ArenaSetupStage, ArenaState, ArenaVariant, ArenaChest).
ArenaCountdownTask.java: Thêm "s" vào chuỗi trạng thái starting để hiển thị "Starting in &aXs" (thêm đơn vị giây).
adminCommand.java: Cập nhật ArenaManager.createArena sử dụng ArenaVariant.fromKey(mode, type) thay vì constructor trực tiếp.
chestHandle.java: Xóa code không cần thiết: setOpened.
ArenaChest.java: Cập nhật fillChest với random items từ 3-6 (tăng từ 2-5), sử dụng Utils.randomInRange. Cập nhật getChestItems phù hợp cấu trúc mới.
ChestItem.java: Chuyển chance từ double sang int để đơn giản hóa.
Hologram.java: Thêm setCustomNameVisible(!text.isEmpty()) trong setLine và spawnLine để ẩn hologram nếu text rỗng, tránh hiển thị không mong muốn.

🚀 Lợi ích: Tăng tính modular, dễ bảo trì hệ thống chest theo variant và phase, cải thiện hiệu suất ngẫu nhiên items, và fix các vấn đề nhỏ về hiển thị/hologram.

+++ src/main/java/com/HauBaka/object/ChestItem.java
@@ -15,9 +15,9 @@ import java.util.stream.Collectors;
 public class ChestItem {
     private final ItemStack item;
     @Getter
-    private final double chance;
+    private final int chance;
 
-    public ChestItem(ItemStack item, double chance) {
+    public ChestItem(ItemStack item, int chance) {
         this.item = item;
         this.chance = chance;
     }
